requirements.txt
flask
python-docx
regex
gunicorn
